CacheType: convert to uint8
We only need ~2 bits for the types of cache. `int` is 64-bits on 64-bit
platforms, which is significant overkill.

Convert CacheType to a uint8 and move it to the end of the `cache`
struct so it only affects alignment, but not the actual size. (we can
pack in other small values at the end later as well)