consistenthash: keys map & int->uint32
Rename the existing `keys` member in `Map` to keyHashes and replace it
with a string-set (map[string]struct{}) that tracks all the actual keys
that have been inserted.

Also, eliminate a bunch of type-conversions by storing uint32 hashes
(matching the hash-function return value) rather than int.
The one disadvantage of storing []uint32 compared to []int in this case
is that we can't use sort.Int(), and instead must use the less efficient
sort.Slice() (unless we write our own sorting routine or create a named
type that implements sort.Interface). Insertions generally aren't in the
critical path, so it seems fine to trade that off for now.