*: add TTL functionality
TTLs are based on expiration in the LRU - during getting if the value is
expired then it is removed from the LRU and we return nothing. This is
needed to have short-lived values in a distributed system. To achieve
this, expand the `valWithStat` struct with `expire time.Time`. Then,
expand all of the related interfaces:
- Codecs
- `setLRUOnEvicted`
- etc.

Since the HTTP interface returns in the body raw bytes, send the
expiration timestamp with headers. Expand the test in
`http/http_test.go` to test whether the keys are being regenerated after
the timestamp. Copy over a test from mailgun/groupcache for TTLs in the
LRU.