lru: Add TypedLRU
Fork the existing LRU cache implementation and add implement a version
based on an LRU stack with strong-typing (using generics).

This should modestly improve the GC-friendliness, in that it will be
possible to use the lru package without forcing the GC to traverse any
pointers other than the linked-list in the LRU-stack.

Gate the new types on go 1.18+ with build tags so go 1.17 users can
still use the package.