hot/candidate cache: remove damping on key qps
The damping was a bit problematic in that it prevented any hits less
than a second after the initial candidate-cache insertion from having a
non-zero qps calculated.

In this new windowed-QPS type, we specifically avoid including a
time.Time so there are no pointers for the Garbage Collector to scan. As
such, we now measure time relative to a new `baseTime` in the Galaxy. To
facilitate testing, time is now managed by an implementation of the
go-clocks package's Clock interface.

Additionally, this has a few other properties:
 - After some time a value is considered "stale" and reset. (so as not
   to penalize a key that's hung in the candidate cache for hours
   relative to one that's been evicted and brought back)
 - QPS is calculated relative to the current "epoch" for that key (which
   is either the creation time or the last time it was reset due to
   staleness.

To facilitate a simpler Promoter implementation, we add a Hits field
that's guaranteed to be 0 on the first fetch.