Promoter that always inserts previously known keys
A la ARC, any second reference to a key that's present in the candidate
cache gets it inserted into the hotcache. This way a key that gets a
short burst of requests is always added to the hotcache.

The candidate-cache is generally not large enough that anything will
hang out in it for a long time anyway.

After some testing, this promoter is likely to become the default.