consistenthash: resolve hash collisions
When generating the hashring, if there are many segments for an inserted
key, there's a nonzero chance of a collision. Previously, the last
insert won and we'd end up with an extra copy of that hash-function
value in the `keyHashes` slice. Change this so the key that sorts last
wins.