protocodec: introduce CodecV2
google.golang.org/protobuf has been the current protobuf version for a
while. Add an updated protocodec and move the test protobuf package to
support it.

Make the new CodecV2 type generic to eliminate the need to type-assert
when dealing with a known-type.

Include a NewV2 helper to make type-inference a bit nicer.