protocodec: make a few tweaks to allow inlining
protocodec.GalaxyGet is particularly hot, and trivial so it should get
inlined. Make a few tweaks to allow inlining. (bringing the cost down
below the magic 80 number in the go compiler)

Similarly, split up protocodec.backendGetterV2.Get to allow inlining of
the fast path. (Amusingly, this also allows the new `setSlow` to inline,
but that wasn't the goal)